refactor: rename monitoring-stack-operator to observability-operator
This patch addresses Issue #145.

With expansion of the scope of the operator to a more generic
observability operator used of deploying Monitoring, Logging and Tracing
stacks, rename the operator to Observability Operator.

Significant changes other than renaming are
  * create separate directory in deployment for `monitoring` related
    resources like prometheus, alertmanager rules.
  * Move monitoring api to `pkg/apis/monitoring` directory
  * Move monitoring controllers to `pkg/controllers/monitoring` directory
  * Move to quay.io/rhobs/observability-operator image repository
