+++ CODEOWNERS
@@ -1,1 +1,2 @@
 * @rhobs/monitoring-stack-operator-maintainers
+* @rhobs/observability-operator-maintainers
