--- deploy/operator/monitoring-stack-operator-rules.yaml
@@ -1,49 +1,0 @@
-apiVersion: monitoring.coreos.com/v1
-kind: PrometheusRule
-metadata:
-  labels:
-    app.kubernetes.io/component: operator
-    app.kubernetes.io/name: monitoring-stack-operator-rules
-    app.kubernetes.io/part-of: monitoring-stack-operator
-    prometheus: k8s
-    role: alert-rules
-  name: monitoring-stack-operator-rules
-spec:
-  groups:
-  - name: monitoring-stack-operator.rules
-    rules:
-    - alert: Monitoring Stack Operator controller reconcile errors
-      annotations:
-        description: |
-          Monitoring Stack Operator controller - {{ $labels.controller }} fails to reconcile.
-          Inspect the monitoring-stack-operator log for potential root causes.
-        summary: Monitoring Stack Operator controller - {{ $labels.controller }} fails to reconcile
-      expr: |
-        increase(controller_runtime_reconcile_errors_total{job="monitoring-stack-operator"}[15m]) > 0
-      for: 15m
-      labels:
-        severity: warning
-    - alert: Monitoring Stack Operator controller reconcilation takes longer than 10 minutes
-      annotations:
-        description: |
-          Monitoring Stack Operator controller reconcilation takes longer than 10 minutes for the controller - {{ $labels.controller }}.
-          Inspect the monitoring-stack-operator log for potential root causes.
-        summary: Monitoring Stack Operator controller - {{ $labels.controller }} reconcilation takes too long to reconcile
-      expr: |
-        rate(controller_runtime_reconcile_time_seconds_sum{job="monitoring-stack-operator"}[5m]) /
-        rate(controller_runtime_reconcile_time_seconds_count{job="monitoring-stack-operator"}[5m])
-        > 600
-      for: 10m
-      labels:
-        severity: warning
-    - alert: Monitoring Stack Operator controller backlog is not being drained
-      annotations:
-        description: |
-          The backlog of Monitoring Stack Operator controller - {{ $labels.name }} is not getting drained; an indication that reconcile loop may be stuck
-          Inspect the monitoring-stack-operator log for potential root causes.
-        summary: Monitoring Stack Operator controller - {{ $labels.name }} backlog is not being drained.
-      expr: |
-        rate(workqueue_depth{job="monitoring-stack-operator"}[15m]) > 0
-      for: 15m
-      labels:
-        severity: critical
