--- deploy/operator/monitoring-stack-operator-service.yaml
@@ -1,15 +1,0 @@
-kind: Service
-apiVersion: v1
-metadata:
-  name: monitoring-stack-operator
-  labels:
-    app.kubernetes.io/component: operator
-    app.kubernetes.io/name: monitoring-stack-operator
-    app.kubernetes.io/part-of: monitoring-stack-operator
-spec:
-  selector:
-    app.kubernetes.io/name: monitoring-stack-operator
-    app.kubernetes.io/component: operator
-  ports:
-  - name: metrics
-    port: 8080
