+++ deploy/operator/observability-operator-cluster-role.yaml
@@ -4,7 +4,7 @@ apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
   creationTimestamp: null
-  name: monitoring-stack-operator
+  name: observability-operator
 rules:
 - apiGroups:
   - ""
