+++ deploy/operator/observability-operator-service-account.yaml
@@ -1,4 +1,4 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: monitoring-stack-operator-sa
+  name: observability-operator-sa
