+++ deploy/operator/observability-operator-service-monitor.yaml
@@ -1,15 +1,15 @@
 apiVersion: monitoring.coreos.com/v1
 kind: ServiceMonitor
 metadata:
-  name: monitoring-stack-operator
+  name: observability-operator
   labels:
     app.kubernetes.io/component: operator
-    app.kubernetes.io/name: monitoring-stack-operator
-    app.kubernetes.io/part-of: monitoring-stack-operator
+    app.kubernetes.io/name: observability-operator
+    app.kubernetes.io/part-of: observability-operator
 spec:
   selector:
     matchLabels:
-      app.kubernetes.io/name: monitoring-stack-operator
+      app.kubernetes.io/name: observability-operator
       app.kubernetes.io/component: operator
   endpoints:
   - port: metrics
