+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/rhobs/monitoring-stack-operator
+module github.com/rhobs/observability-operator
 
 go 1.17
