+++ hack/olm/subscription.yaml
@@ -3,13 +3,13 @@ apiVersion: operators.coreos.com/v1alpha1
 kind: Subscription
 metadata:
   labels:
-    operators.coreos.com/monitoring-stack-operator.openshift-operators: ""
-  name: monitoring-stack-operator
+    operators.coreos.com/observability-operator.openshift-operators: ""
+  name: observability-operator
   namespace: openshift-operators
 spec:
   channel: stable
   installPlanApproval: Automatic
-  name: monitoring-stack-operator
+  name: observability-operator
   source: monitoring-operators
   sourceNamespace: openshift-marketplace
-  startingCSV: monitoring-stack-operator.v0.0.6
+  startingCSV: observability-operator.v0.0.10
