+++ jsonnet/components/alertmanager/alertmanager-rules.jsonnet
@@ -8,11 +8,11 @@ local rules = (
 {
   _commonLabels:: {
     'app.kubernetes.io/component': 'operator',
-    'app.kubernetes.io/name': 'monitoring-stack-operator-alertmanager-rules',
-    'app.kubernetes.io/part-of': 'monitoring-stack-operator',
+    'app.kubernetes.io/name': 'observability-operator-alertmanager-rules',
+    'app.kubernetes.io/part-of': 'observability-operator',
     prometheus: 'k8s',
     role: 'alert-rules',
   },
 
-  rule: $.k.prometheusrule.new('monitoring-stack-operator-alertmanager-rules', $._commonLabels, rules),
+  rule: $.k.prometheusrule.new('observability-operator-alertmanager-rules', $._commonLabels, rules),
 }
