+++ jsonnet/components/prometheus-operator/prometheus-operator-rules.jsonnet
@@ -3,11 +3,11 @@ local rules = import 'github.com/prometheus-operator/prometheus-operator/jsonnet
 {
   _commonLabels:: {
     'app.kubernetes.io/component': 'operator',
-    'app.kubernetes.io/name': 'monitoring-stack-operator-prometheus-operator-rules',
-    'app.kubernetes.io/part-of': 'monitoring-stack-operator',
+    'app.kubernetes.io/name': 'observability-operator-prometheus-operator-rules',
+    'app.kubernetes.io/part-of': 'observability-operator',
     prometheus: 'k8s',
     role: 'alert-rules',
   },
 
-  rule: $.k.prometheusrule.new('monitoring-stack-operator-prometheus-operator-rules', $._commonLabels, rules),
+  rule: $.k.prometheusrule.new('observability-operator-prometheus-operator-rules', $._commonLabels, rules),
 }
