+++ jsonnet/components/prometheus/prometheus-rules.jsonnet
@@ -8,11 +8,11 @@ local rules = (
 {
   _commonLabels:: {
     'app.kubernetes.io/component': 'operator',
-    'app.kubernetes.io/name': 'monitoring-stack-operator-prometheus-rules',
-    'app.kubernetes.io/part-of': 'monitoring-stack-operator',
+    'app.kubernetes.io/name': 'observability-operator-prometheus-rules',
+    'app.kubernetes.io/part-of': 'observability-operator',
     prometheus: 'k8s',
     role: 'alert-rules',
   },
 
-  rule: $.k.prometheusrule.new('monitoring-stack-operator-prometheus-rules', $._commonLabels, rules),
+  rule: $.k.prometheusrule.new('observability-operator-prometheus-rules', $._commonLabels, rules),
 }
