+++ pkg/apis/monitoring/v1alpha1/register.go
