+++ pkg/controllers/monitoring/monitoring-stack/components_test.go
