+++ pkg/operator/scheme.go
@@ -1,7 +1,7 @@
 package operator
 
 import (
-	rhobsv1alpha1 "github.com/rhobs/monitoring-stack-operator/pkg/apis/v1alpha1"
+	rhobsv1alpha1 "github.com/rhobs/observability-operator/pkg/apis/monitoring/v1alpha1"
 
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
